refactor: decode_ddr_bytes を parser モジュールに集約
commands/import.rs とtests/integration_ddr.rs に重複していた
デコードロジックを parser::decode_ddr_bytes として一本化。
戻り値を Result<String, String> から Result<String, ParseError> に変更し、
commands/error.rs に From<ParseError> for CommandError を追加。

Closes #27
